Bug: Fix how CartoPy projections are handled in plot_model_data.py
As explained in this StackedOverflow post: https://stackoverflow.com/questions/42237802/plotting-projected-data-in-other-projectons-using-cartopy, the 'projection' and 'transform' keywords do not need to use the same projection. 'projection' refers to the projection used for plotting the data whereas 'transform' is the projection of the underlying data itself. If (lat, lon) coordinates are used, then 'transform' should be PlateCarree(). Thus, the code here was updated so that 'transform' is always PlateCarree() and the user is free to select the projection for 'projection'.