Bug: Fix how GSI diag files are read in gsi_fcts.py
The bug resulted in each observation being read 3 times when converting the GSI diag DataSets to DataFrames. This occurred b/c the DataSet had 2 dimensions (nobs and Bias_Correction_Terms_arr_dim). The latter dimension had a length of 3, which is why each variable was read 3 times. To fix this bug, the Bias_Correction_Terms variable is dropped prior to converting to a DataFrame.